Refactor: images 폴더 분류
- PostImages, PrevieImages, ProfileImage
