+++ frontend/src/components/images/PostImages/style.ts
