+++ frontend/src/components/images/PreviewImages/style.ts
