+++ frontend/src/components/images/ProfileImage/index.tsx
