+++ frontend/src/components/images/ProfileImage/style.ts
